requirements.txt
# Core
numpy>=1.24.0
pillow>=10.0.0
torch>=2.0.0
transformers>=4.38.0

# Frontend
streamlit>=1.28.0

opencv-python-headless>=4.8.0
pytesseract>=0.3.10
easyocr>=1.7.0
